backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
firebase-admin==6.4.0
google-cloud-aiplatform>=1.39.0
google-cloud-firestore==2.13.1
langchain>=0.2.0
langchain-google-vertexai>=0.1.0
langgraph>=0.1.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
